Narrow GPUI history BDD support modules
Compile constructors and direct entity access only in the history
integration-test binaries that need them. This removes artificial
unused-code anchors while preserving genuine detection of unused helpers.
