+++ tests/gpui_history_bdd/support.rs
@@ -1,9 +1,7 @@
 //! Durable GPUI handles shared by the history BDD step libraries.
 use gauss::ui::Phase0Shell;
-use gpui::{AnyWindowHandle, Entity, TestAppContext, VisualContext, VisualTestContext};
+use gpui::{AnyWindowHandle, Entity, TestAppContext, VisualTestContext};
 use test_support::{TestSupportError, TestSupportResult};
-
-use crate::common::{ensure_initial_draw, init_test_app};
 
 /// Handles that may safely outlive an individual BDD step.
 ///
@@ -26,60 +24,11 @@ use crate::common::{ensure_initial_draw, init_test_app};
 /// ```
 #[derive(Clone)]
 pub struct DurableShell {
-    entity: Entity<Phase0Shell>,
-    window: AnyWindowHandle,
+    pub(crate) entity: Entity<Phase0Shell>,
+    pub(crate) window: AnyWindowHandle,
 }
 
 impl DurableShell {
-    /// Create a normal Phase 0 shell for interaction-driven scenarios.
-    ///
-    /// The returned handle owns the window and entity references needed by
-    /// subsequent BDD steps.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// # use gpui::TestAppContext;
-    /// # use crate::history_bdd_support::DurableShell;
-    /// # fn given_shell(cx: &mut TestAppContext) -> DurableShell {
-    /// DurableShell::open(cx)
-    /// # }
-    /// ```
-    pub fn open(cx: &mut TestAppContext) -> Self {
-        init_test_app(cx);
-        let (entity, visual_cx) = cx.add_window_view(|_window, view_cx| Phase0Shell::new(view_cx));
-        ensure_initial_draw(visual_cx);
-        Self {
-            entity,
-            window: visual_cx.window_handle(),
-        }
-    }
-
-    /// Create a Phase 0 shell exposing history test seams.
-    ///
-    /// The returned handle is ready for history-specific step definitions to
-    /// use through [`Self::with_visual`].
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// # use gpui::TestAppContext;
-    /// # use crate::history_bdd_support::DurableShell;
-    /// # fn given_history_shell(cx: &mut TestAppContext) -> DurableShell {
-    /// DurableShell::open_for_tests(cx)
-    /// # }
-    /// ```
-    pub fn open_for_tests(cx: &mut TestAppContext) -> Self {
-        init_test_app(cx);
-        let (entity, visual_cx) =
-            cx.add_window_view(|_window, view_cx| Phase0Shell::new_for_tests(view_cx));
-        ensure_initial_draw(visual_cx);
-        Self {
-            entity,
-            window: visual_cx.window_handle(),
-        }
-    }
-
     /// Rebuild a visual context for the current harness step.
     ///
     /// The closure receives a context valid for the duration of the call and
@@ -104,30 +53,7 @@ impl DurableShell {
         let mut visual_cx = VisualTestContext::from_window(self.window, cx);
         f(&mut visual_cx, &self.entity)
     }
-
-    /// Read state directly after an interaction that closes the visual context.
-    ///
-    /// The reference can be passed to a read-only helper while the current
-    /// step owns the `TestAppContext`.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// # use gauss::ui::Phase0Shell;
-    /// # use gpui::{Entity, TestAppContext};
-    /// # use crate::history_bdd_support::DurableShell;
-    /// # fn read_entity(shell: &DurableShell, _cx: &mut TestAppContext) {
-    /// let _: &Entity<Phase0Shell> = shell.entity();
-    /// # }
-    /// ```
-    pub const fn entity(&self) -> &Entity<Phase0Shell> {
-        &self.entity
-    }
 }
-
-const _: fn(&mut TestAppContext) -> DurableShell = DurableShell::open;
-const _: fn(&mut TestAppContext) -> DurableShell = DurableShell::open_for_tests;
-const _: for<'a> fn(&'a DurableShell) -> &'a Entity<Phase0Shell> = DurableShell::entity;
 
 /// Report a missing scenario value without panicking in a step.
 ///
