+++ tests/gpui_history_command_grouping_undo.rs
@@ -8,6 +8,8 @@ mod common;
 
 #[path = "gpui_history_bdd/support.rs"]
 mod history_bdd_support;
+#[path = "gpui_history_bdd/support_open_for_tests.rs"]
+mod history_bdd_support_open_for_tests;
 
 use common::{read_document, read_history_len, simulate_document_undo};
 use gauss::model::history::HistoryError;
