+++ tests/gpui_history_multi_shape_drag_undo.rs
@@ -5,6 +5,8 @@ mod common;
 
 #[path = "gpui_history_bdd/support.rs"]
 mod history_bdd_support;
+#[path = "gpui_history_bdd/support_open.rs"]
+mod history_bdd_support_open;
 
 #[path = "gpui_history_bdd/multi_shape.rs"]
 mod multi_shape;
