+++ tests/gpui_history_reorder_undo.rs
@@ -8,6 +8,8 @@
 mod common;
 #[path = "gpui_history_bdd/support.rs"]
 mod history_bdd_support;
+#[path = "gpui_history_bdd/support_open.rs"]
+mod history_bdd_support_open;
 #[path = "gpui_history_bdd/reorder.rs"]
 mod reorder;
